% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{COCA.calibration}
\alias{COCA.calibration}
\title{Calibrate COCA parameters}
\usage{
COCA.calibration(
  case,
  n.stage1 = 24,
  n.stage2 = c(),
  dosage.ctrl = c(A = 0, B = 0),
  dosage.singleA = 0,
  dosage.singleB = 0,
  dosage.comb = list(A = c(), B = c()),
  eff.null,
  eff.alt.SOC,
  eff.alt.A = 0,
  eff.alt.B = 0,
  eff.alt.AB,
  period.effect = c(0.1, 0.2, 0.3),
  alpha.level = 0.1,
  alpha.max = 0.2,
  fsr.level = 0.05,
  power.target = 0.9,
  tsr.target = 0.8,
  prior.sample = 1e+06,
  seed = 123,
  n.simu = 10000
)
}
\arguments{
\item{case}{Trial type for stage 2. \code{case = 1} for 4-arm trial comparing AB vs. A vs. B vs. SOC; \code{case = 2} for 3-arm trial comparing AB vs. A (or B) vs. SOC; \code{case = 3} for 2-arm trial comparing AB vs. SOC.}

\item{n.stage1}{Sample size for stage 1}

\item{n.stage2}{A numeric vector of candidate stage 2 sample sizes to evaluate during calibration.}

\item{dosage.ctrl}{Dosage level of the control arm for stage 2. For an SOC control, use \code{c(A = 0, B = 0)}. If one of the single agents is used as the SOC (e.g., drug A 300 mg), use \code{c(A = 300, B = 0)}.}

\item{dosage.singleA}{Dosage level of drug A in the single arm for stage 2.}

\item{dosage.singleB}{Dosage level of drug B in the single arm for stage 2.}

\item{dosage.comb}{A named list specifying the dosage levels of drugs A and B across combination arms in stage 1. For example, \code{list(A = c(300, 300, 200), B = c(300, 200, 300))} defines three dose combinations:
dose1 = (A = 300, B = 300), dose2 = (A = 300, B = 200), and dose3 = (A = 200, B = 300).
All dosage values can be on any scale but must use the same scale across \code{dosage.singleA}, \code{dosage.singleB}, and \code{dosage.comb}.}

\item{eff.null}{Unpromising efficacy rate (\eqn{\widetilde{q}_1}) in the global null hypothesis}

\item{eff.alt.SOC}{Efficacy rate of the SOC arm (\eqn{\widetilde{q}_{SOC}}) in the alternative hypothesis}

\item{eff.alt.A}{Efficacy rate of single arm A (\eqn{\widetilde{q}_{A}}) in the alternative hypothesis. This argument is ignored if arm A is not included.}

\item{eff.alt.B}{Efficacy rate of single arm B (\eqn{\widetilde{q}_{B}}) in the alternative hypothesis. This argument is ignored if arm B is not included.}

\item{eff.alt.AB}{Efficacy rate of combination arm AB (\eqn{\widetilde{q}_{AB}}) in the alternative hypothesis.}

\item{period.effect}{Vector of possible period effects}

\item{alpha.level}{Type I error level (\eqn{\alpha_0}) under no period effect assumption}

\item{alpha.max}{Maximum type I error level (\eqn{\alpha_{\max}}) under non-zero period effect}

\item{fsr.level}{False success rate level (\eqn{\gamma_0}) under no period effect assumption}

\item{power.target}{Target power (\eqn{\phi_0}) under no period effect assumption}

\item{tsr.target}{Target true success rate (\eqn{\zeta_0}) under no period effect assumption}

\item{prior.sample}{Number of prior draws in each simulation}

\item{seed}{Random seed}

\item{n.simu}{Number of simulation replicates}

\item{n.comb.dose}{Number of combination arms in stage 1}
}
\value{
A named list with the following elements:
\describe{
\item{full.results}{A tibble data frame containing the calibrated cutoffs \eqn{C_{e1}} and \eqn{c_0}, and the corresponding power and type I error rates for all tried sample sizes.}
\item{optimal.config}{A tibble data frame with the same metrics for the first sample size meeting the target criteria.}
}
}
\description{
Computes the optimal design cutoffs (\eqn{C_{e1}} and \eqn{c_0}) and the corresponding power across a range of stage 2 sample sizes for a given configuration.
}
\details{
For each candidate value in \code{n.stage2}, the function computes the calibrated cutoffs \eqn{C_{e1}} and \eqn{c_0}, along with power and type I error rates. It returns the full results across all sample sizes considered, as well as the optimal configuration that first meets the target power and validity criteria.
}
\examples{

\donttest{
COCA.calibration(
  case = 1, n.stage1 = 24, n.stage2 = seq(20, 26, 2),
  dosage.ctrl = c(A = 0, B = 0), dosage.singleA = 300, dosage.singleB = 300,
  dosage.comb = list(A = c(300, 300, 200), B = c(300, 200, 300)),
  eff.null = 0.25, eff.alt.SOC = 0.25, eff.alt.A = 0.35,
  eff.alt.B = 0.35, eff.alt.AB = 0.55, period.effect = c(0.1, 0.2, 0.3),
  alpha.level = 0.10, alpha.max = 0.20, fsr.level = 0.05,
  power.target = 0.90, tsr.target = 0.80,
  prior.sample = 1e4, seed = 123, n.simu = 1000
)
}

}
