% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coca.R
\name{COCA.getOC}
\alias{COCA.getOC}
\title{COCA Simulation}
\usage{
COCA.getOC(
  case = 1,
  n.stage1 = 24,
  n.stage2,
  Ce,
  c0,
  tox.SOC = 0.1,
  eff.SOC = 0.25,
  tox.A = 0.25,
  tox.B = 0.15,
  eff.A = 0.25,
  eff.B = 0.25,
  tox.AB = c(0.3, 0.3, 0.15),
  eff.AB = c(0.25, 0.25, 0.25),
  tox.upper = 0.35,
  eff.lower = 0.25,
  Cs = 0.85,
  C.f1 = 0.9,
  C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100),
  rho = 0.2,
  period.effect = 0,
  n.simu = 10,
  seed = 123
)
}
\arguments{
\item{case}{Trial type for stage 2. \code{case = 1} for 4-arm trial comparing AB vs. A vs. B vs. SOC; \code{case = 2} for 3-arm trial comparing AB vs. A (or B) vs. SOC; \code{case = 3} for 2-arm trial comparing AB vs. SOC.}

\item{n.stage1}{Sample size for stage 1}

\item{n.stage2}{Sample size for stage 2}

\item{Ce}{Design cutoff, obtained using \code{COCA.calibration} fucntion.}

\item{c0}{Design cutoff, obtained using \code{COCA.calibration} fucntion.}

\item{tox.SOC, tox.A, tox.B}{True toxicity probabilities for SOC, arm A, and arm B.}

\item{eff.SOC, eff.A, eff.B}{True efficacy probabilities for SOC, arm A, and arm B.}

\item{tox.AB}{A vector of true toxicity probabilities for all combination doses being tested in the trial.}

\item{eff.AB}{A vector of true efficacy probabilities for all combination doses being tested in the trial.}

\item{tox.upper}{Highest acceptable toxicity rate (\eqn{\phi_{T}})}

\item{eff.lower}{Lowest acceptable efficacy rate (\eqn{\phi_{E}})}

\item{Cs}{Probability cutoff in stage 1}

\item{C.f1, C.f2}{Probability cutoffs in stage 2}

\item{utility.score}{Vector of utility score: \code{(b1, b2, b3, b4)} represents the utility for (toxicity, no efficacy), (no toxicity, no efficacy), (toxicity, efficacy), and (no toxicity, efficacy), respectively.}

\item{rho}{Correlation between toxicity and efficacy}

\item{period.effect}{Period effect}

\item{n.simu}{Number of simulation replicates}

\item{seed}{Random seed}
}
\value{
Returns the operating characteristics of stage 1 (selection and expected sample size) and stage 2 (power, GP, SR, OSR, and expected sample size).
}
\description{
COCA Simulation
}
\examples{
COCA.getOC(n.stage2 = 26, Ce = 0.8983, c0 = 0.7)
}
