% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{computeAUC}
\alias{computeAUC}
\title{Compute Area Under the Curve (AUC)}
\usage{
computeAUC(consensusMatrix)
}
\arguments{
\item{consensusMatrix}{Consensus matrix, output of the
"coca::consensusCluster" function.}
}
\value{
This function returns the area under the curve.
}
\description{
This function allows to compute the area under the curve of the empirical
distribution function of a consensus matrix as described in Monti et al.
(2003), Section 3.3.1.
}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk}
}
\keyword{internal}
