% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus-cluster.R
\name{consensusCluster}
\alias{consensusCluster}
\title{Consensus clustering with k-means}
\usage{
consensusCluster(data, K, B = 100, pItem = 0.8, clMethod = "km",
  distHC = "euclidean")
}
\arguments{
\item{data}{N X P data matrix}

\item{K}{number of clusters}

\item{B}{of iterations}

\item{pItem}{proportion of items sampled at each iteration}

\item{clMethod}{clustering method}

\item{distHC}{Distance used for hierarchical clustering. Can be "pearson" (for 1 - Pearson
correlation), "spearman" (for 1- Spearman correlation), or any of the distances provided in
stats::dist() (i.e. "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski").}
}
\value{
The output is a consensus matrix, that is a symmetric matrix where the element in
position (i,j) corresponds to the proportion of times that items i and j have been clustered
together.
}
\description{
This function allows to perform consensus clustering using the k-means clustering algorithm,
for a fixed number of clusters. We consider the number of clusters K to be fixed.
}
\examples{
## Load one dataset with 300 observations, 2 variables, 6 clusters
data <- as.matrix(read.csv(system.file("extdata", "dataset1.csv", package = "coca"),
row.names = 1))

## Compute consensus clustering with K=6 clusters
cm <- consensusCluster(data, K = 6)
}
\references{
Monti, S., Tamayo, P., Mesirov, J. and Golub, T., 2003. Consensus clustering:
a resampling-based method for class discovery and visualization of gene expression microarray
data. Machine learning, 52(1-2), pp.91-118.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
