% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dose_standardize.R
\name{DoseStandardize}
\alias{DoseStandardize}
\title{Standardize Drug Dosage}
\usage{
DoseStandardize(d)
}
\arguments{
\item{d}{Dose vectors for all considered doses}
}
\value{
Standardized doses
}
\description{
Standardize Drug Dosage
}
\examples{
dosage_level <- c(300, 200, 0)
DoseStandardize(dosage_level)
}
