% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand-moc.R
\name{expandMOC}
\alias{expandMOC}
\title{Expand matrix of clusters}
\usage{
expandMOC(clLabels, datasetNames = NULL)
}
\arguments{
\item{clLabels}{Matrix of cluster labels of size N x M.}

\item{datasetNames}{Vector of cluster names of length M. Default is null.}
}
\value{
The output is a list containing the matrix of clusters `moc`, a vector containing
the dataset indicator `datasetIndicator` and a vector of expanded `datasetNames` of
length sum(K).
}
\description{
Expand matrix of cluster labels into matrix of clusters
}
\examples{
## Load data
data <- list()
data[[1]] <- as.matrix(read.csv(system.file("extdata", "dataset1.csv",
package = "coca"), row.names = 1))
data[[2]] <- as.matrix(read.csv(system.file("extdata", "dataset2.csv",
package = "coca"), row.names = 1))
data[[3]] <- as.matrix(read.csv(system.file("extdata", "dataset3.csv",
package = "coca"), row.names = 1))

## Build matrix of clusters
outputBuildMOC <- buildMOC(data, M = 3, K = 6, distances = "cor")

## Extract matrix of clusters
clLabels <- outputBuildMOC$clLabels

## Impute missing values
outputFillMOC <- fillMOC(clLabels, fullData = TRUE, data = data)

## Replace matrix of cluster labels with new (full) one
clLabels <- outputFillMOC$fullClLabels

## Expand matrix of cluster labels into matrix of clusters
outputExpandMOC <- expandMOC(clLabels)
clLabels <- outputExpandMOC$clLabels
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
