% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill-moc.R
\name{fillMOC}
\alias{fillMOC}
\title{Fill Matrix-Of-Clusters}
\usage{
fillMOC(clLabels)
}
\arguments{
\item{clLabels}{N X M matrix containing cluster labels. Element (n,m) contains the cluster label for element}
}
\value{
The output is the same matrix of clusters, where NAs have been replaced by their estimates, where possible.
}
\description{
This function fills in a matrix of clusters that contains NAs, by estimating the missing cluster labels
based on the available ones.
}
\references{
The Cancer Genome Atlas, 2012. Comprehensive molecular portraits of human breast tumours. Nature,
487(7407), pp.61–70.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
