% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_example.R
\name{open_example}
\alias{open_example}
\title{Open Example R Script}
\usage{
open_example(name = "example_code.R")
}
\arguments{
\item{name}{File name in \verb{inst/examples/}}
}
\description{
Opens an example R script in the editor.
}
\examples{
\donttest{
open_example("alt_designs.R")

}

}
