% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{plotDeltaAUC}
\alias{plotDeltaAUC}
\title{Plot area under the curve}
\usage{
plotDeltaAUC(deltaAUC, chosenK = NULL, fileName)
}
\arguments{
\item{chosenK}{Chosen number of clusters. If specified, a vertical line is plotted in
correspondance of the indicated value. Default is NULL.}

\item{fileName}{name of the png file}

\item{areaUnderTheCurve}{vector of the average silhouette for K from 2 to some value maxK}
}
\description{
Plot area under the curve for different values of K
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
