% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-moc.R
\name{plotMOC}
\alias{plotMOC}
\title{Plot Matrix-Of-Clusters}
\usage{
plotMOC(moc, datasetIndicator, datasetNames, annotations = NA,
  clr = FALSE, clc = FALSE, save = FALSE, fileName = "moc.pdf")
}
\arguments{
\item{moc}{Matrix-Of-Clusters of size N x sumK.}

\item{datasetIndicator}{Vector containing}

\item{datasetNames}{Vector containing the names of the datasets to which each column of
labels corresponds. If NULL, datasetNames will be the same as datasetIndicator.
Default is NULL.}

\item{annotations}{Vector or matrix of annotations of size N x a where a is the
number of different annotations.}

\item{clr}{Cluster rows. Default is FALSE.}

\item{clc}{Cluster columns. Default is FALSE.}

\item{save}{Boolean. If TRUE, plot is saved as a pdf file.}

\item{fileName}{File name. Default is "moc.pdf".}
}
\description{
This function creates a matrix of clusters, starting from a list of heterogeneous datasets.
}
\references{
The Cancer Genome Atlas, 2012. Comprehensive molecular portraits of human breast
tumours. Nature, 487(7407), pp.61–70.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
