% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plot silhouette}
\usage{
plotSilhouette(sil, chosenK = NULL, fileName)
}
\arguments{
\item{sil}{vector of the average silhouette for K from 2 to some value maxK}

\item{chosenK}{Chosen number of clusters. If specified, a vertical line is plotted in
correspondance of the indicated value. Default is NULL.}

\item{fileName}{name of the png file}
}
\description{
Plot average silhouette.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
