% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{plotWidestGap}
\alias{plotWidestGap}
\title{Plot widest gap}
\usage{
plotWidestGap(widestGap, chosenK = NULL, fileName)
}
\arguments{
\item{widestGap}{Vector of widest gap values for K from 2 to some value maxK.}

\item{chosenK}{Chosen number of clusters. If specified, a vertical line is
plotted in correspondence of the indicated value. Default is NULL.}

\item{fileName}{Name of the png file.}
}
\value{
invisible(0)
}
\description{
Plot widest gap for different values of K.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
