% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-moc.R
\name{buildMOC}
\alias{buildMOC}
\title{Build Matrix-Of-Clusters}
\usage{
buildMOC(data, M, K, methods, full = FALSE)
}
\arguments{
\item{data}{List of M datasets, each of size N X P_m, m = 1, ..., M.}

\item{M}{number of datasets.}

\item{K}{Vector containing the number of clusters in each dataset. If given an integer
instead of a vector it is assumed that each dataset has the same number of clusters}

\item{methods}{A vector of strings containing the names of the clustering methods to be
used to cluster the observations in each dataset}

\item{full}{Boolean. If TRUE, if there are any missing observations in one or more datasets,
the corresponding cluster labels will be estimated through generalised linear models on the
basis of the available labels.}
}
\value{
The output is the Matrix-Of-Clusters (MOC). It is a binary matrix of size N x sum(K)
where element (n,k) contains a 1 if observation n belongs to the corresponding cluster,
0 otherwise. It also returns a vector called datasetIndicator of length sum(K) in which
each element is the number of the dataset to which the cluster belongs.
}
\description{
This function creates a matrix of clusters, starting from a list of heterogeneous datasets.
}
\references{
The Cancer Genome Atlas, 2012. Comprehensive molecular portraits of human
breast tumours. Nature, 487(7407), pp.61–70.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
