% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{chooseKusingAUC}
\alias{chooseKusingAUC}
\title{Choose number of clusters based on AUC}
\usage{
chooseKusingAUC(areaUnderTheCurve, savePNG, fileName)
}
\arguments{
\item{areaUnderTheCurve}{Vector of length maxK-1 containing the area
under the curve of the empirical distribution function of the
consensus matrices obtained with K varying from 2 to maxK.}
}
\value{
This function returns a list containing `deltaAUC`, a vector of
        length maxK-1 where element i is  the area under the curve for
        K = i+1 minus the area under the curve for K = i (for i = 2 this
        is simply the area under the curve for K = i), and `K` the lowest
        among the values of K that are chosen by the algorithm.
}
\description{
This function allows to choose the number of clusters in a dataset
based on the area under the curve of the empirical distribution
function of a consensus matrix, calculated for different (consecutive)
cluster numbers, as explained in the article by Monti et al. (2003),
Section 3.3.1.
}
