% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{COCA.calibration}
\alias{COCA.calibration}
\title{COCA Calibration}
\usage{
COCA.calibration(
  case,
  n.stage2,
  lower.eff = 0.25,
  period.effect = c(0.1, 0.2, 0.3),
  alpha.level = 0.1,
  alpha.max = 0.2,
  seed = 123,
  n.simu = 20
)
}
\arguments{
\item{case}{Trial type for stage 2. \code{case = 1} for 4-arm trial comparing AB vs. A vs. B vs. SOC; \code{case = 2} for 3-arm trial comparing AB vs. A (or B) vs. SOC; \code{case = 3} for 2-arm trial comparing AB vs. SOC.}

\item{n.stage2}{Sample size for stage 2}

\item{lower.eff}{Lowest acceptable efficacy rate}

\item{period.effect}{Vector of possible period effects}

\item{alpha.level}{Type I error level (\eqn{\alpha_0}) under no period effect assumption}

\item{alpha.max}{Maximum type I error level (\eqn{\alpha_{\max}}) under non-zero period effect}

\item{seed}{Random seed}

\item{n.simu}{Number of simulation replicates}
}
\value{
BCI's
}
\description{
COCA Calibration
}
\examples{

# To calibrate for a specific sample size candidate, run:
COCA.calibration(case = 1, n.stage2 = 20, lower.eff = 0.25, period.effect = c(0.1, 0.2, 0.3), seed = 123)

# For a grid search, try:
n.stage2 <- 10:20
for(i in 1:length(n.stage2)){
  if(i == 1){
    output <- COCA.calibration(case = 1, n.stage2 = n.stage2[i], lower.eff = 0.25, period.effect = c(0.1, 0.2, 0.3), seed = 123)
  } else {
    output.tmp <- COCA.calibration(case = 1, n.stage2 = n.stage2[i], lower.eff = 0.25, period.effect = c(0.1, 0.2, 0.3), seed = 123)
    output <- bind_rows(output, output.tmp)
  }

}

}
