% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coca.R
\name{COCA.getOC}
\alias{COCA.getOC}
\title{COCA Simulation}
\usage{
COCA.getOC(
  case = 1,
  n.stage1 = 24,
  n.stage2,
  Ce,
  c0,
  tox.SOC = 0.1,
  eff.SOC = 0.25,
  tox.A = 0.25,
  tox.B = 0.15,
  eff.A = 0.25,
  eff.B = 0.25,
  tox.AB = c(0.3, 0.3, 0.15),
  eff.AB.s1 = c(0.25, 0.25, 0.25),
  eff.AB.s2 = c(0.25, 0.25, 0.25),
  tox.upper = 0.35,
  eff.lower = 0.25,
  Cs = 0.85,
  C.f1 = 0.9,
  C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100),
  rho = 0.2,
  n.simu = 10,
  seed = 123
)
}
\arguments{
\item{case}{Trial type for stage 2. \code{case = 1} for 4-arm trial comparing AB vs. A vs. B vs. SOC; \code{case = 2} for 3-arm trial comparing AB vs. A (or B) vs. SOC; \code{case = 3} for 2-arm trial comparing AB vs. SOC.}

\item{n.stage1}{Sample size for stage 1}

\item{n.stage2}{Sample size for stage 2}

\item{Ce, c0}{Design cutoffs, obtained using the \code{COCA.calibration} function.}

\item{tox.SOC, tox.A, tox.B}{True toxicity probabilities for SOC, arm A, and arm B.}

\item{eff.SOC, eff.A, eff.B}{True efficacy probabilities for SOC, arm A, and arm B.}

\item{tox.AB}{A vector of true toxicity probabilities for all combination doses being tested in the trial.}

\item{eff.AB.s1}{A vector of true efficacy probabilities for all combination doses in stage 1.}

\item{eff.AB.s2}{A vector of true efficacy probabilities for all combination doses in stage 2.}

\item{tox.upper}{Highest acceptable toxicity rate (\eqn{\phi_{T}})}

\item{eff.lower}{Lowest acceptable efficacy rate (\eqn{\phi_{E}})}

\item{Cs}{Probability cutoff in stage 1}

\item{C.f1, C.f2}{Probability cutoffs in stage 2}

\item{utility.score}{Vector of utility score: \code{c(b1, b2, b3, b4)} represents the utility for (toxicity, no efficacy), (no toxicity, no efficacy), (toxicity, efficacy), and (no toxicity, efficacy), respectively.}

\item{rho}{Correlation between toxicity and efficacy}

\item{n.simu}{Number of simulation replicates. The default value \code{n.simu = 10} is used for illustration purposes and is small to reduce computation time. For more accurate results, consider using a larger value, such as 5000.}

\item{seed}{Random seed}

\item{period.effect}{Period effect}
}
\value{
Returns the operating characteristics of stage 1 (selection and expected sample size) and stage 2 (power, GP, SR, OSR, and expected sample size).
}
\description{
COCA Simulation
}
\examples{
# \code{n.simu = 20} is used for illustration purposes. For more accurate results,
# consider using a larger value for \code{n.simu}, such as 5000.
# Scenario 1 (period effect = 0)
\donttest{
COCA.getOC(
  n.stage2 = 26, Ce = 0.8983, c0 = 0.7,
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.25, eff.B = 0.25, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.25, 0.25, 0.25), eff.AB.s2 = c(0.25, 0.25, 0.25),
  n.simu = 20
)}
# Scenario 1 (period effect = 0.2)
\donttest{
COCA.getOC(
  n.stage2 = 26, Ce = 0.8983, c0 = 0.7,
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.25, eff.B = 0.25, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.45, 0.45, 0.45), eff.AB.s2 = c(0.25, 0.25, 0.25),
  n.simu = 20
)}
# Scenario 2 (period effect = 0)
\donttest{
COCA.getOC(
  n.stage2 = 26, Ce = 0.8983, c0 = 0.7,
  tox.SOC = 0.10, eff.SOC = 0.25,  = 0.25, tox.B = 0.15,
  eff.A = 0.35, eff.B = 0.35, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.55, 0.55, 0.55), eff.AB.s2 = c(0.55, 0.55, 0.55),
  n.simu = 20
)}
# Scenario 2 (period effect = 0.2)
\donttest{
COCA.getOC(
  n.stage2 = 26, Ce = 0.8983, c0 = 0.7,
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.35, eff.B = 0.35, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.75, 0.75, 0.75), eff.AB.s2 = c(0.55, 0.55, 0.55),
  n.simu = 20
)}
}
