% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_coca.R
\name{COCA.getOC}
\alias{COCA.getOC}
\title{Run the COCA Design}
\usage{
COCA.getOC(
  case,
  n.stage1 = 24,
  n.stage2,
  Ce,
  c0,
  nlook.stage1 = 2,
  nlook.stage2 = 2,
  dosage.ctrl = c(A = 0, B = 0),
  dosage.singleA = 0,
  dosage.singleB = 0,
  dosage.comb = list(A = c(), B = c()),
  tox.SOC,
  eff.SOC,
  tox.A = 0,
  tox.B = 0,
  eff.A = 0,
  eff.B = 0,
  tox.AB = c(),
  eff.AB.s1 = c(),
  eff.AB.s2 = c(),
  tox.isomat,
  tox.upper,
  eff.lower,
  Cs = 0.85,
  Ct = 0.9,
  C.f1 = 0.9,
  C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100),
  rho = 0.2,
  prior.sample = 1e+06,
  seed = 123,
  n.simu = 5000
)
}
\arguments{
\item{case}{Trial type for stage 2. \code{case = 1} for 4-arm trial comparing AB vs. A vs. B vs. SOC; \code{case = 2} for 3-arm trial comparing AB vs. A (or B) vs. SOC; \code{case = 3} for 2-arm trial comparing AB vs. SOC.}

\item{n.stage1}{Sample size for stage 1}

\item{n.stage2}{Sample size for stage 2}

\item{Ce, c0}{Design cutoffs, obtained using the \code{COCA.calibration} function.}

\item{nlook.stage1}{The total number of planned analyses in stage 1, including all interim analyses and the final analysis (e.g., \code{nlook.stage1 = 2} for one interim and one final analysis in stage 2).}

\item{nlook.stage2}{The total number of planned analyses in stage 2, including all interim analyses and the final analysis (e.g., \code{nlook.stage2 = 2} for one interim and one final analysis in stage 2).}

\item{dosage.ctrl}{Dosage level of the control arm for stage 2. For an SOC control, use \code{c(A = 0, B = 0)}.If one of the single agents is used as the SOC (e.g., drug A 300 mg), use \code{c(A = 300, B = 0)}.}

\item{dosage.singleA}{Dosage level of drug A in the single arm for stage 2.}

\item{dosage.singleB}{Dosage level of drug B in the single arm for stage 2.}

\item{dosage.comb}{A named list specifying the dosage levels of drugs A and B across combination arms in stage 1.
For example, \code{list(A = c(300, 300, 200), B = c(300, 200, 300))} defines three dose combinations:
dose1 = (A = 300, B = 300), dose2 = (A = 300, B = 200), and dose3 = (A = 200, B = 300).
All dosage values can be on any scale but must use the same scale across \code{dosage.singleA}, \code{dosage.singleB}, and \code{dosage.comb}.}

\item{tox.SOC, tox.A, tox.B}{True toxicity probabilities for SOC, arm A, and arm B.}

\item{eff.SOC, eff.A, eff.B}{True efficacy probabilities for SOC, arm A, and arm B.}

\item{tox.AB}{A vector of true toxicity probabilities for all combination doses being tested in the trial.}

\item{eff.AB.s1}{A vector of true efficacy probabilities for all combination doses in stage 1.}

\item{eff.AB.s2}{A vector of true efficacy probabilities for all combination doses in stage 2.}

\item{tox.isomat}{Matrix with 2 columns that contains isotonicity conditions for the toxicity order among the J combination doses. The format should follow the structure expected by the \code{activeSet()} function in the \code{isotone} package. For details, refer to the \href{https://cran.r-project.org/web/packages/isotone/index.html}{isotone package documentation}.}

\item{tox.upper}{Highest acceptable toxicity rate (\eqn{\phi_{T}})}

\item{eff.lower}{Lowest acceptable efficacy rate (\eqn{\phi_{E}})}

\item{Cs, Ct}{Probability cutoffs in stage 1}

\item{C.f1, C.f2}{Probability cutoffs in stage 2}

\item{utility.score}{Vector of utility score: \code{c(b1, b2, b3, b4)} represents the utility for (toxicity, no efficacy), (no toxicity, no efficacy), (toxicity, efficacy), and (no toxicity, efficacy), respectively.}

\item{rho}{Correlation between toxicity and efficacy}

\item{prior.sample}{Number of prior draws used to fit the Bayesian model in each simulation}

\item{seed}{Random seed}

\item{n.simu}{Number of simulation replicates. The default value \code{n.simu = 10} is used for illustration purposes and is small to reduce computation time. For more accurate results, consider using a larger value, such as 5000.}
}
\value{
Returns the operating characteristics of stage 1 (selection and expected sample size) and stage 2 (power, GP, SR, OSR, and expected sample size).
}
\description{
This function uses \code{activeSet()} from the \code{isotone} package to perform isotonic regression.
}
\details{
Run the COCA design and get its operating characteristics.
}
\examples{
# \code{prior.sample = 1e5} is used for illustration purposes. For more accurate results,
# consider using a larger value for \code{prior.sample}, such as 10^6.

# Scenario 1 (period effect = 0)
\donttest{
COCA.getOC(
  case = 1, n.stage1 = 24, n.stage2 = 26, Ce = 0.9152, c0 = 0.66,
  dosage.ctrl = c(A = 0, B = 0), dosage.singleA = 300, dosage.singleB = 300,
  dosage.comb = list(A = c(300, 300, 200), B = c(300, 200, 300)),
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.25, eff.B = 0.25, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.25, 0.25, 0.25), eff.AB.s2 = c(0.25, 0.25, 0.25),
  tox.isomat = matrix(c(2, 1, 3, 1), byrow = TRUE, nrow = 2),
  tox.upper = 0.35, eff.lower = 0.25, Cs = 0.85, Ct = 0.9, C.f1 = 0.9, C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100), rho = 0.2, prior.sample = 1e5, n.simu = 5000
)
}

# Scenario 1 (period effect = 0.2)
\donttest{
COCA.getOC(
  case = 1, n.stage1 = 24, n.stage2 = 26, Ce = 0.9152, c0 = 0.66,
  dosage.ctrl = c(A = 0, B = 0), dosage.singleA = 300, dosage.singleB = 300,
  dosage.comb = list(A = c(300, 300, 200), B = c(300, 200, 300)),
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.25, eff.B = 0.25, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.45, 0.45, 0.45), eff.AB.s2 = c(0.25, 0.25, 0.25),
  tox.isomat = matrix(c(2, 1, 3, 1), byrow = TRUE, nrow = 2),
  tox.upper = 0.35, eff.lower = 0.25, Cs = 0.85, Ct = 0.9, C.f1 = 0.9, C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100), rho = 0.2, prior.sample = 1e5, n.simu = 5000
)
}

# Scenario 2 (period effect = 0)
\donttest{
COCA.getOC(
  case = 1, n.stage1 = 24, n.stage2 = 26, Ce = 0.9152, c0 = 0.66,
  dosage.ctrl = c(A = 0, B = 0), dosage.singleA = 300, dosage.singleB = 300,
  dosage.comb = list(A = c(300, 300, 200), B = c(300, 200, 300)),
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.35, eff.B = 0.35, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.55, 0.55, 0.55), eff.AB.s2 = c(0.55, 0.55, 0.55),
  tox.isomat = matrix(c(2, 1, 3, 1), byrow = TRUE, nrow = 2),
  tox.upper = 0.35, eff.lower = 0.25, Cs = 0.85, Ct = 0.9, C.f1 = 0.9, C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100), rho = 0.2, prior.sample = 1e5, n.simu = 5000
)
}

# Scenario 2 (period effect = 0.2)
\donttest{
COCA.getOC(
  case = 1, n.stage1 = 24, n.stage2 = 26, Ce = 0.9152, c0 = 0.66,
  dosage.ctrl = c(A = 0, B = 0), dosage.singleA = 300, dosage.singleB = 300,
  dosage.comb = list(A = c(300, 300, 200), B = c(300, 200, 300)),
  tox.SOC = 0.10, eff.SOC = 0.25, tox.A = 0.25, tox.B = 0.15,
  eff.A = 0.35, eff.B = 0.35, tox.AB = c(0.30, 0.30, 0.15),
  eff.AB.s1 = c(0.75, 0.75, 0.75), eff.AB.s2 = c(0.55, 0.55, 0.55),
  tox.isomat = matrix(c(2, 1, 3, 1), byrow = TRUE, nrow = 2),
  tox.upper = 0.35, eff.lower = 0.25, Cs = 0.85, Ct = 0.9, C.f1 = 0.9, C.f2 = 0.9,
  utility.score = c(0, 60, 40, 100), rho = 0.2, prior.sample = 1e5, n.simu = 5000
)
}
}
