% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coca.R
\name{coca}
\alias{coca}
\title{Cluster-Of-Clusters Analysis}
\usage{
coca(MOC, K, B = 1000, pItem = 0.8, hclustMethod = "average")
}
\arguments{
\item{MOC}{N X C data matrix, where C is the total number of clusters considered.}

\item{K}{number of clusters.}

\item{B}{number of iterations of the Consensus Cluster step.}

\item{pItem}{proportion of items sampled at each iteration of the Consensus Cluster step.}

\item{hclustMethod}{method to be used by the hclust function for the hierarchical clustering step.}
}
\value{
The output is a consensus matrix, that is a symmetric matrix where the element in position (i,j) corresponds to
the proportion of times that items i and j have been clustered together and a vector of cluster labels.
}
\description{
This function allows to do Cluster-Of-Clusters-Analysis on a binary matrix where each column is a clustering of the data,
each row corresponds to a data point and the element in position (i,j) is equal to 1 if data point i belongs to cluster
j, 0 otherwise.
}
\examples{
## Load data
data <- list()
data[[1]] <- as.matrix(read.csv(system.file("extdata", "dataset1.csv",
package = "coca"), row.names = 1))
data[[2]] <- as.matrix(read.csv(system.file("extdata", "dataset2.csv",
package = "coca"), row.names = 1))
data[[3]] <- as.matrix(read.csv(system.file("extdata", "dataset3.csv",
package = "coca"), row.names = 1))

## Set constants
n_datasets <- 3
n_clusters <- 6
N <- 300
true_labels <- as.matrix(read.csv(system.file("extdata", "cluster_labels.csv",
package = "coca"), row.names = 1))

## Fill label matrix with clusterings found with the k-means clustering algorithm
labelMatrix <- array(NA, c(n_clusters, N, n_datasets))
for(i in 1:n_datasets){
  output <- kmeans(data[[i]], n_clusters)
  for(k in 1:n_clusters){
    labelMatrix[k,,i] <- (output$cluster==k)
  }
}
# Convert label matrix from logic to numeric matrix
labelMatrix <- labelMatrix*1
}
\references{
The Cancer Genome Atlas, 2012. Comprehensive molecular portraits of human breast tumours. Nature,
487(7407), pp.61–70.
}
\author{
Alessandra Cabassi \email{ac2051@cam.ac.uk}
}
