% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{plotDunn2s}
\alias{plotDunn2s}
\title{Plot Dunn's alternative index}
\usage{
plotDunn2s(dunns, chosenK = NULL, fileName)
}
\arguments{
\item{dunns}{Vector of widest gap values for K from 2 to some value maxK.}

\item{chosenK}{Chosen number of clusters. If specified, a vertical line is
plotted in correspondence of the indicated value. Default is NULL.}

\item{fileName}{Name of the png file.}
}
\description{
Plot Dunn's alternative index (minimum average dissimilarity between two
cluster divided by maximum average within cluster dissimilarity) for
different values of K.
}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk}
}
\keyword{internal}
