% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silhouette.R
\name{plotSilhouette}
\alias{plotSilhouette}
\title{Plot silhouette}
\usage{
plotSilhouette(sil, chosenK = NULL, fileName)
}
\arguments{
\item{sil}{vector of the average silhouette for K from 2 to some value maxK}

\item{chosenK}{Chosen number of clusters. If specified, a vertical line is
plotted in correspondence of the indicated value. Default is NULL.}

\item{fileName}{Name of the png file.}
}
\description{
Plot average silhouette for different values of K.
}
\author{
Alessandra Cabassi \email{alessandra.cabassi@mrc-bsu.cam.ac.uk}
}
\keyword{internal}
